move function to navigation.py
