+++ src/traffic/algorithms/navigation.py
@@ -26,6 +26,7 @@ import numpy as np
 import pandas as pd
 from shapely.geometry import LineString, MultiLineString, Point, Polygon, base
 
+from ..core.distance import minimal_angular_difference
 from ..core.iterator import flight_iterator
 from ..core.time import deltalike, to_timedelta
 
@@ -652,6 +653,132 @@ class NavigationFeatures:
             )
             if result is not None:
                 yield result
+
+    def get_toff_runway(
+        self,
+        airport: Union[str, Airport],
+        threshold_alt: float = 1500,
+        min_groundspeed_kts: float = 30,
+        min_vert_rate_ftmin: float = 257,
+        max_bearing_deg: float = 10,
+        max_dist_nm: float = 5,
+    ) -> Optional[str]:
+        """
+        Determines the taking-off runway of a flight based on its trajectory.
+
+        Parameters:
+            flight (Flight): The flight data object containing flight data.
+            airport: The ICAO code of the airport. Defaults to None.
+            threshold_alt (float): Altitude threshold above airport altitude.
+            min_groundspeed_kts (float): Minimum groundspeed to consider.
+            min_vert_rate_ftmin (float): Minimum vertical rate to consider.
+            maximum_bearing_deg (float): Maximum bearing difference to consider.
+            max_dist_nm (float): Maximum distance from the airport to consider.
+
+        Returns:
+            Optional[str]: The name of the closest runway, or None if not found.
+
+        Raises:
+            KeyError: If the airport code is not found in the airports data.
+            ValueError: If no rway available or data is empty after filtering.
+        """
+        from traffic.data import airports
+
+        airport = airport if isinstance(airport, Airport) else airports[airport]
+        if not self.takeoff_from(airport):
+            return None
+        alt_max = airport.altitude + threshold_alt
+        if airport.runways is None:
+            return None
+        runways = airport.runways.data
+        runways_names = runways.name
+
+        self = cast("Flight", self)
+
+        filtered_flight = self.distance(airport).query(
+            f"distance < {max_dist_nm}"
+        )
+        if filtered_flight is None or filtered_flight.data.empty:
+            return None
+        if "geoaltitude" in filtered_flight.data.columns:
+            query_str = (
+                f"geoaltitude < {alt_max} and "
+                f"vertical_rate > {min_vert_rate_ftmin} and "
+                f"groundspeed > {min_groundspeed_kts}"
+            )
+        else:
+            query_str = (
+                f"altitude < {alt_max} and "
+                f"vertical_rate > {min_vert_rate_ftmin} and "
+                f"groundspeed > {min_groundspeed_kts}"
+            )
+        filtered_flight = filtered_flight.query(query_str)
+        if (
+            filtered_flight is None
+            or filtered_flight.data.empty
+            or len(filtered_flight.data) < 4
+        ):
+            return None
+
+        # Check for parallel runways with suffixes L, R, or C
+        has_parallel_runway = runways_names.str.contains(r"[LRC]").any()
+        runway_bearings = runways.bearing
+
+        median_track = filtered_flight.data["track"].median()
+        closest_runway: Optional[str] = None
+
+        if not has_parallel_runway:
+            # Find the runway with the bearing closest to the median track
+            bearing_diff = runway_bearings.apply(
+                lambda x: minimal_angular_difference(x, median_track)
+            )
+            closest_index = bearing_diff.idxmin()
+            closest_runway = runways.name.iloc[closest_index]
+        else:
+            # Round runway bearings to the nearest 5 degrees
+            rounded_bearings = (runway_bearings / 5).round() * 5
+            rounded_bearings = (
+                rounded_bearings % 360
+            )  # Ensure bearings stay within 0-359
+
+            # Find the bearing closest to the median track
+            bearing_diff = rounded_bearings.apply(
+                lambda x: minimal_angular_difference(x, median_track)
+            )
+
+            # Identify all runways where bearing diff is less than 10 deg
+
+            candidate_runways = runways.loc[
+                bearing_diff[bearing_diff < max_bearing_deg].index
+            ]
+            if candidate_runways.empty:
+                return None
+
+            # Calculate distance from flight trajectory to each candidate runway
+            try:
+                flight_ls = filtered_flight.linestring
+                rways_ls = airport.runways.shape
+
+                closest = None
+                for rway in rways_ls.geoms:
+                    if closest is None:
+                        closest = rway
+                        continue
+                    if rway.distance(flight_ls) < closest.distance(flight_ls):
+                        closest = rway
+                if closest is None:
+                    return None
+                lon_1, _, lon_2, _ = closest.bounds
+                # candidate_runways =
+                # lat/lon that is closest to the closest bounds
+                eps = 1 / 1000
+                closest_runway = candidate_runways.query(
+                    f"(abs(longitude-{lon_1})<{eps}) or "
+                    "(abs(longitude-{lon_2})<{eps})"
+                )["name"].iloc[0]
+            except AttributeError:
+                return None
+        return closest_runway
 
     # -- Special operations --
